SI-9666: Use inline group names in Regex
Delegate `Match group name` to the underlying `matcher`.

If that fails, try explicit group names as a fall back.

No attempt is made to correlate inline and explicit names.

In the following case, either name is accepted:

```
new Regex("a(?<Bar>b*)c", "Bee")
```

But if names are reversed, the error is undetected:

```
new Regex("a(?<Bee>b*)(?<Bar>c)", "Bar", "Bee")
```

Throw IllegalArg on bad group name to be consistent with Java.