Extract API for each class separately
The ExtractAPI collects API for each class (including inner classes)
separately. For example,

class A {
  class B {
    def foo: Int = 123
  }
}

Is represented as:

// className = A
class A {
  class B
}
// className = A.B
class A.B {
  def foo: Int
}

The signature of an inner classes is represented twice:

  1. as a member of an outer class
  2. as a standalone class

The first representation is needed so dependencies by inheritance are
invalidated properly even if they don't depend explicitly on the inner
class.

The class-based-inheritance test has been expanded to show that changes
to an API of an inner class do not affect classes inheriting from an outer
class.

Rewritten from sbt/zinc@934eaa206e95cf69c441cb5cefb6175f5e6549b1