Update STARR to freshly bootstrapped build
As built in https://scala-ci.typesafe.com/view/scala-2.12.x/job/scala-2.12.x-integrate-bootstrap/326/console

To pick this up in PR validation, I've configured SBT to use the extra
resolver. Once we gain some more confidence about the new STARR candidate,
we'll tag it and have the bootstrap script publish it to Sonatype, at
which time we can drop this change.