Skip module docs in bootstrap script when STARR_REF is used
This is a second attempt at making this script operate when
making a binary incompatible change to the compiler, such as
the current effort to change the trait encoding.

I have just disabled the docs when STARR_REF is provided,
rather than disabling them in the first pass as I tried in
c4fc2fd. scala-dev/#89 showed that we can't defer docs to
the second pass, as the second pass isn't always run.