bundle a newer Akka actors jar
might as well let people have the current stuff. I asked the
Akka folks and nobody saw a reason not to upgrade.  this
version of Akka is the latest in the 2.3.x series, but is
also some months old, so no stability worries.

(this would have been included in 2.11.8 if I had thought of
it in time)