Encapsulate the AST encoding of arbitrary super calls
Created a factory method for an AST shape that is used in
a number of places to symbolically bind to a particular
super method without needed to specify the qualifier of
the `Super` tree (which is too limiting, as it only allows
you to refer to direct parents.)

It was too awkward to encapsulate this further, so I've
left the pattern match in the backend as it was, and added
a comment to like it to `SelectSuper`.

I also found a similar tree shape created in Delambdafy,
that is better expressed with an existing tree creation
factory method, mkSuperInit.