Skip module docs in first round of bootstrap when STARR_REF is used
This is a second attempt at making this script operate when
making a binary incompatible change to the compiler, such as
the current effort to change the trait encoding.

For builds that don't define STARR_REF, we continue to build docs
in the first pass, as there will only be one pass for non-release
builds.

A previous attempt than disabling them in the first pass for all
builds (c4fc2fd) triggered scala-dev/#89.