Add get() tests to OpenHashMap, reduce timing artifacts.
In order to get a better exploration of the variance of tests in a
limited time, I've reduced the number of measurement iterations and
increased the number of forks.  By sight, the measurement iterations
seemed pretty consistent within a trial, whereas they would vary
widely on occasional forks.

I extended testing down to 50-entry maps, to explore the rise in
service times that I was seeing at small scale.  This is probably a
timing artifact, from too-short invocations, since I'm using
@Level.Invocation in the put() tests.  To fix that, I enlarged the
unit of testing, by creating multiple, sometimes thousands, of maps
for the invocation to fill.  This has also changed the test from
filling a previously-filled map, to filling a new, but sufficiently
sized map.  The put()/remove() test now performs much worse (on a
more realistic scenario).

This also adds a couple tests for calling get() against a map that's
been filled only with put()s, or with a mix of put() and remove().