Don't cache the hash codes in OpenHashMap.
The original author did some superficial exploration of whether this
change would improve performance, but seemed to think it was slower.
After some intensive benchmarking, it seems to me to be slightly
faster, but not statistically significant.  More useful is that it
uses about 10% less memory to not cache the hash codes.