Remove the Option usage from OpenHashMap.
Replacing the Option wrappers on values, with a Byte array of state
indicators, and removing the HashEntry trait,
- cuts memory usage by 10%, by not storing the Some instances, and
  another 10% by not storing the HashEntry field
- speeds put() & remove() for maps w/ >100k entries, by as much as
  50-100%, apparently by delaying memory effects

But it also
- slows put() by up to 10% for maps w/ 100 to 20k entries
- slows get() by up to 20% for maps constructed w/ put() & remove()

put() is slowed by adding reference to the separate Byte array.  get()
is slowed by adding a Some wrapper in the call; this could be avoided
in an optimized getOrElse().  Both of these are probably mitigated at
larger sizes by the reduced memory usage.

Some foreach() calls also were rewritten to while-loops for better code
generation.

This commit also modifies a test for a change in compiler mangling of
private member variable names.  Not sure why this wasn't caught in a
previous commit.