SI-9684 Use only JavaConverters
Reacting to the trending popularity of JavaConverters, the
other mechanisms are deprecated.

To be complete, the deprecated implicits could be duplicated and
renamed; it would be handy if one could mark them
deprecatedImplicitly to emit a warning if they were used
implicitly.

The public traits bearing the deprecated implicits are marked
deprecatedInheritance.