Expect exactly FunctionN for function literal
We used to allow subtypes of FunctionN as the expected
type of a function literal, which (only) works out if
an implicit conversion closes the gap between the
inferred FunctionN type and the expected subtype.

This leads to bizarre error messages, and now we have
SAM conversion, it's redundant. In retrospect, I don't
think the allowance for type slack was according to spec.

This fixes scala/scala-dev#138.

TODO:
  - reopen SI-6221
  - ensure Spark and the spores library can use sam conversion