SI-9762 Update to JLine 2.14.3
Cherry-picked 365ac03 (and bumped from 2.14.1 to 2.14.3)

Motivated by the improvements to multi-byte character handling.

Screencast showing the reported bug is fixed:

  http://g.recordit.co/ie1Z367NUl.gif

Here's the changelog since JLine 2.12.1:

  https://github.com/jline/jline2/compare/jline-2.12.1...jline-2.14.3

I needed to disable a new, on-by-default feature in JLine so that
it didn't add a " " after completing the token `equals` in
`foo.equa<TAB>`.