SD-110 Emit self-synchronized trait methods with ACC_SYNCHRONIZED
Avoids the bloat of spelling out the details in the method body.

If we elect to emit trait methods in static methods with an explicit self
parameter (which is under consideration to avoid headaches with using
invokespecial to methods in non-direct parent interfaces), we would
need to revert this change.