Revert pruning of redundant interfaces
This reverts the fix for SI-5278 made in 7a99c03da
and later refined.

Direct parents enable use of `invokespecial` in the JVM,
so are not redundant. The motivating use case was the
proliferation of the `ScalaObject` parent in Android.
We no longer include that interfaces as a parent of
any class, and Scala 2.12 has left Android behind for
the time being due to use of Java 8 facilities.