SI-9781 Don't convert erroneous expression to assignment
`convertToAssignment` is triggered on a type error but it doesn't seem
to really care what the error is as long as the expression can be
converted to an assignment expression.

This patch fixes that by checking whether the qualifier of the selection
contains any errors before deciding to continue with the conversion.