SI-9104 Autodetect raw pastage
If `-raw` is not supplied explicitly to REPL `:paste`,
see if the code text starts with `package` keyword or
else see if it parses to a named package (to cope with
leading commentary). In that case, take it as raw.

But parse only on suspect comment slash.

It's only worth parsing for a package if there's a chance
that package keyword is buried behind comments.

Small refactors to the `paste` object.