Simplify erasure + mixin
Remove some old, obsolete & untested hacks from ExplicitOuter.
Added a test for one of them to show this is now fine.

There are a lot of `makeNotPrivate` invocations sprinkled around
the codebase. Lets see if we can centralize the ones dealing
with trait methods that need implementations in the phase that emits them.

For example Fields (accessors for fields/modules) or SuperAccessors.