Don't pass -opt to starr / locker build in the bootstrap script
Renaming -Yopt to -opt revealed that we're passing the flag when
building the locker (and optionally the starr) version. This is not
necessary: when building the next stage with the flag enabled, the same
optimizations are performed no matter if the current stage was built
with the flag or not.