Avoid tree sharing with substituteThis
The underlying transformer has a by-name parameter for the
to provide the `to` tree, but this was strict in the layers
of API above.

Tree sharing is frowned upon in general as it leads to cross
talk when, e.g., the erasure typechecker mutates the
`tpe` field of the shared tree in different context.